Updated Battle System A Lot
Updates include:

-You can now enter battles from the overworld screen!
--Instead of taking you to the inside of the house, the "enter square"
will instead take you to a battle screen.
--HOW IT WORKS
--Due to Unity being finicky about keeping data between scenes, we have
a ungodly abomination of static variables.
--Now we have a class called a "Battle Loader"
---You can pass in strings of enemy names in "Turn Based Battle
Manager", and when the battle manager is initialized, it will intantiate
enemies based on the list of strings that are passed in

-Fixed some Battle GUI formatting

TODO
-----------
Implement Loading the Player
---You can see the structure already. There will be a basic class that
contains the player data, and since you can set a gameobject's component
to an instance of a script, I had to work around it by creating an new
script and then "pasting" the og script into the compent-ized version.
--Implement saving/updating data between battles

VIDEO GAMES